+++ R/credentials_byo_oauth2.R
@@ -2,6 +2,10 @@
 #'
 #' @description
 #' This function does very little when called directly with a token:
+#'   * If input has class `request`, i.e. it is a token that has been prepared
+#'     with [httr::config()], the `auth_token` component is extracted. For
+#'     example, such input could be produced by `googledrive::drive_token()`
+#'     or `bigrquery::bq_token()`.
 #'   * Checks that the input appears to be a Google OAuth token, based on
 #'     the embedded `oauth_endpoint`.
 #'   * Refreshes the token, if it's refreshable.
@@ -12,6 +16,18 @@
 #' gargle does not support incremental authorization. The main point of
 #' `credentials_byo_oauth2()` is to allow `token_fetch()` (and packages that
 #' wrap it) to accomodate a "bring your own token" workflow.
+#'
+#' This also makes it possible to obtain a token with one package and then
+#' register it for use with another package. For example, the default scope
+#' requested by googledrive is also sufficient for operations available in
+#' googlesheets4. You could use a shared token like so:
+#' ```
+#' library(googledrive)
+#' library(googlesheets4)
+#' drive_auth(email = "jane_doe@example.com")
+#' sheets_auth(token = drive_token())
+#' # work with both packages freely now
+#' ```
 #'
 #' @inheritParams token_fetch
 #' @param token A token with class [Token2.0][httr::Token-class].
@@ -27,6 +43,9 @@
 #' }
 credentials_byo_oauth2 <- function(scopes = NULL, token, ...) {
   cat_line("trying credentials_byo_oauth()")
+  if (inherits(token, "request")) {
+    token <- token$auth_token
+  }
   stopifnot(inherits(token, "Token2.0"))
 
   if (!is.null(scopes)) {